pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=9.2.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "streamlit>=1.44.0",
]
